# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Documentação Erp",
    "description": "Documentação",
    "version": "1.0.0"
  },
  "host": "http://localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/fornecedores": {
      "get": {
        "tags": [
          "Fornecedor"
        ],
        "description": "Endpoint para listar Fornecedores.",
        "produces": [],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "perPage",
            "in": "query",
            "type": "string"
          },
          {
            "name": "q",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Fornecedor"
            },
            "description": "Listando Fornecedores"
          }
        }
      },
      "post": {
        "tags": [
          "Fornecedor"
        ],
        "description": "Endpoint para cadastrar Fornecedores.",
        "parameters": [
          {
            "name": "Entidade",
            "in": "body",
            "description": "Informações do Fornecedor.",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Fornecedor"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/registerSuccess"
            },
            "description": "Adicionado com sucesso."
          },
          "409": {
            "schema": {
              "$ref": "#/definitions/emailNo"
            },
            "description": "Fornecedor com e-mail ${email} já está registado"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "registerSuccess": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Registado com Sucesso!"
        }
      }
    },
...
    "emailNo": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "E-mail informado Já existe!"
        }
      }
    },
...
    "Fornecedor": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "example": "Nzila"
        },
        "email": {
          "type": "string",
          "example": "nzila@yahoo.com"
        },
        "contacto1": {
          "type": "string",
          "example": "123456789"
        },
        "contacto2": {
          "type": "string",
          "example": "821458963"
        },
        "endereco1": {
          "type": "string",
          "example": "Maputo"
        },
        "nuit": {
          "type": "number",
          "example": 13484849
        },
        "nuel": {
          "type": "string",
          "example": "nuel"
        },
        "plafond": {
          "type": "number",
          "example": 1.2
        },
        "taxa_isento": {
          "type": "number",
          "example": 1
        },
        "motivo": {
          "type": "string",
          "example": "motivo"
        },
        "desconto_fixo": {
          "type": "number",
          "example": 19
        },
        "endereco": {
          "type": "string",
          "example": "Matola"
        },
        "empresa": {
          "type": "number",
          "example": 1
        },
        "data_added": {
          "type": "number",
          "example": 2000
        },
        "data_updated": {
          "type": "number",
          "example": 2000
        },
        "usuario_added": {
          "type": "number",
          "example": 1
        },
        "usuario_updated": {
          "type": "number",
          "example": 1
        }
      }
    },
...
  },
...
}